Reduce unnecessary Mongo polling.
Specifically, updates and removes that specify a specific _id, and all inserts,
should not require live queries that specify a specific different _id to poll.

This involves changing the InvalidationCrossbar matching semantics. I suspect
that a future version of the InvalidationCrossbar may want the old semantics, if
the notifications it is receiving are more oplog-driven than command-driven, but
this will work for now.

The implementation passes the test suite, but it would be nice to:

 - add unit tests for the crossbar
 - add mongo_livedata_tests tests testing that things still poll
 - do a manual test to show that unnecessary polls don't happen
 - benchmark a typical app which does single-ID writes and observes